+++ src/minified-generated-full-src.js
@@ -969,14 +969,17 @@ define('minified', function() {
 		function push(obj, prop, value) {
 			(obj[prop] = (obj[prop] || [])).push(value);
 		}
-		if (isFunction(eventSpec)) 
+		if (isFunction(subSelector)) 
+			return this['on'](null, null, subSelector, eventSpec, handler);
+		else if (isFunction(eventSpec)) 
 			return this['on'](null, subSelector, eventSpec, handler, bubbleSelector);
 		else if (isString(args)) 
 			return this['on'](subSelector, eventSpec, handler, null, args);
 		else
 			return this['each'](function(baseElement, index) {
 				flexiEach(subSelector ? dollarRaw(subSelector, baseElement) : baseElement, function(el) {
-					flexiEach(eventSpec.split(/\s/), function(namePrefixed) {
+					flexiEach(toString(eventSpec).split(/\s/), function(evSpec) {
+						var namePrefixed = evSpec || (/form/i.test(el['tagName']) ? 'submit' : 'click');
 						var name = replace(namePrefixed, /[?|]/);
 						var miniHandler = createEventHandler(handler, el, args,	index, replace(namePrefixed, /[^?|]/g), bubbleSelector && getFilterFunc(bubbleSelector, el));
 		
@@ -1000,17 +1003,20 @@ define('minified', function() {
 	// @condend ie8compatibility 
 	// @condblock !ie8compatibility 
 	function onNonCompat(subSelector, eventSpec, handler, args, bubbleSelector) {
-		if (isFunction(eventSpec)) 
-			return this['on'](null, subSelector, eventSpec, handler, bubbleSelector);
+		if (isFunction(subSelector)) 
+			return this['on'](null, null, subSelector, eventSpec, handler);
+		else if (isFunction(eventSpec)) 
+			return this['on'](null, subSelector, eventSpec, handler, args);
 		else if (isString(args)) 
 			return this['on'](subSelector, eventSpec, handler, null, args);
 		else
 			return this['each'](function(baseElement, index) {
 				flexiEach(subSelector ? dollarRaw(subSelector, baseElement) : baseElement, function(registeredOn) {
-					flexiEach(eventSpec.split(/\s/), function(namePrefixed) {
+					flexiEach(toString(eventSpec).split(/\s/), function(evSpec) {
+						var namePrefixed = evSpec || (/form/i.test(registeredOn['tagName']) ? 'submit' : 'click');
 						var name = replace(namePrefixed, /[?|]/);
 						var prefix = replace(namePrefixed, /[^?|]/g);
-						
+
 						var miniHandler = function(event, triggerOriginalTarget) {
 							var stop;
 							var match = !bubbleSelector;
@@ -3919,6 +3925,7 @@ define('minified', function() {
 	 * @requires dollar each
 	 * @configurable default
 	 * @name .on()
+	 * @syntax list.on(eventHandler)
 	 * @syntax list.on(names, eventHandler)
 	 * @syntax list.on(selector names, eventHandler)
 	 * @syntax list.on(names, customFunc, args)
@@ -3967,6 +3974,11 @@ define('minified', function() {
 	 * $('#myButton').on('click', setStatus, ['running']);
 	 * </pre>
 	 *
+	 * As 'click' is the default event for buttons, you can also omit it here:
+	 * <pre>
+	 * $('#myButton').on(setStatus, ['running']);
+	 * </pre>
+	 *
 	 * @example Adds two handlers on an input field. The event names are prefixed with '|' and thus keep their original behaviour: 
 	 * <pre>
 	 * $('#myInput').on('|keypress |keydown', function() {
@@ -3998,11 +4010,13 @@ define('minified', function() {
 	 * @param selector optional a selector string for ##dollar#$() to register the event only on those children of the list elements that
 	 *                match the selector. 
 	 *                Supports all valid parameters for ##dollar#$() except functions.            
-	 * @param names the space-separated names of the events to register for, e.g. 'click'. Case-sensitive. The 'on' prefix in front of 
+	 * @param names optional the space-separated names of the events to register for, e.g. 'click'. Case-sensitive. The 'on' prefix in front of 
 	 *             the name must not used. You can register the handler for more than one event by specifying several 
 	 *             space-separated event names. If the name is prefixed
 	 *             with '|' (pipe), the event will be passed through and the event's default actions will be executed by the browser. 
-	 *             If the name is prefixed with '?', the event will only be passed through if the handler returns <var>true</var>. 
+	 *             If the name is prefixed with '?', the event will only be passed through if the handler returns <var>true</var>.
+	 *             If you omit the parameter, Minified will chose the default event type for each element. The default is 'submit' for
+	 *             forms and 'click' for everything else. 
 	 * @param eventHandler the callback <code>function(event, index, selectedIndex)</code> to invoke when the event has been triggered:
 	 * 		  <dl>
  	 *             <dt>event</dt><dd>The original DOM event object.</dd>